Unit tests for Rust code
This PR includes unit tests for the bulk of the newly written Rust code. In addition, it fixes a number of bugs which were identified during the process of writing the tests. Any python tests which were broken by recent changes have been brought back to a working state.
While there is still a significant backlog of work, the routes generated by Rust are now comparable in quality to those which were being created by the old Python implementation. Once containerised, this project should be in a good place for an initial release.